Add caching of v8 Strings used for property lookup
Before we were creating new Strings all the time for property lookup
from C.  This caused us to spend a lot of time in String::New and
created garbage that needed to be collected.

This patch takes all constant strings and puts them into Persistent
handles to cache them.

This was done with a script - it is not added in this patch, ask Tom
about it.