restore copyGameFiles functionality
Restore copyGameFiles functionality to match copyFiles and the docs. Now files
listed in module config.json `copyGameFiles` array will be copied from the
game folder to a matching path in the build folder. This roughly matches
the `copyFiles` functionality - in the future, it would be better to support
a map of source to destination files instead of forcing the files to start
and end with the same path.