Pass memory warning to texture manager
We need to consider android system's memory warning seriously,
otherwise the system will kill the app.

This issue is pretty frequent when user is having a device which
currently has very low free memory. Passing the warning to native will
help us to remove textures or even half size it.

This is already being done in iOS.