reset critical memory flag on low trim value
We don't need to send the same trim value to native, if onTrimMemory is
called multiple time by the OS. It will try to reduce memory limit every
time and max texture memory will reach a point where it won't be able to
hold all the texture we want to load in the screen.

Also, I found that when memory situation becomes better, OS is calling
onTrimMemory with lower values. We were handling it by half sizing,
which is wrong.